remove invalid tests
move functions out of utils
address review comments
revert lockfile changes
